requirements.txt
pymorphy2~=0.8
PyYAML
python-telegram-bot
requests~=2.22.0
pymongo~=3.10.1